+++ git_cat-ainahack/utils/audioUtils.ts
@@ -1,0 +1,88 @@
+import { Platform } from 'react-native';
+import { Audio } from 'expo-av';
+import * as FileSystem from 'expo-file-system';
+
+export const convertToWav = async (audioBlob: Blob): Promise<Blob> => {
+  const audioContext = new (window.AudioContext || window.AudioContext)();
+  const arrayBuffer = await audioBlob.arrayBuffer();
+  const audioBuffer = await audioContext.decodeAudioData(arrayBuffer);
+  
+  const numberOfChannels = audioBuffer.numberOfChannels;
+  const length = audioBuffer.length;
+  const sampleRate = audioBuffer.sampleRate;
+  const wavBuffer = audioContext.createBuffer(numberOfChannels, length, sampleRate);
+  
+  for (let channel = 0; channel < numberOfChannels; channel++) {
+    wavBuffer.copyToChannel(audioBuffer.getChannelData(channel), channel);
+  }
+  
+  const wavBlob = await new Promise<Blob>(resolve => {
+    const offlineContext = new OfflineAudioContext(numberOfChannels, length, sampleRate);
+    const source = offlineContext.createBufferSource();
+    source.buffer = wavBuffer;
+    source.connect(offlineContext.destination);
+    source.start();
+    
+    offlineContext.startRendering().then(renderedBuffer => {
+      const wav = new Blob([exportWAV(renderedBuffer)], { type: 'audio/wav' });
+      resolve(wav);
+    });
+  });
+  
+  return wavBlob;
+};
+
+function exportWAV(audioBuffer: AudioBuffer): ArrayBuffer {
+  const interleaved = interleaveChannels(audioBuffer);
+  const dataView = encodeWAV(interleaved, audioBuffer.sampleRate, audioBuffer.numberOfChannels);
+  return dataView.buffer;
+}
+
+function interleaveChannels(audioBuffer: AudioBuffer): Float32Array {
+  const numberOfChannels = audioBuffer.numberOfChannels;
+  const length = audioBuffer.length * numberOfChannels;
+  const result = new Float32Array(length);
+
+  for (let i = 0; i < audioBuffer.length; i++) {
+    for (let channel = 0; channel < numberOfChannels; channel++) {
+      result[i * numberOfChannels + channel] = audioBuffer.getChannelData(channel)[i];
+    }
+  }
+  return result;
+}
+
+function encodeWAV(samples: Float32Array, sampleRate: number, numChannels: number): DataView {
+  const buffer = new ArrayBuffer(44 + samples.length * 2);
+  const view = new DataView(buffer);
+
+  writeString(view, 0, 'RIFF');
+  view.setUint32(4, 36 + samples.length * 2, true);
+  writeString(view, 8, 'WAVE');
+  writeString(view, 12, 'fmt ');
+  view.setUint32(16, 16, true);
+  view.setUint16(20, 1, true);
+  view.setUint16(22, numChannels, true);
+  view.setUint32(24, sampleRate, true);
+  view.setUint32(28, sampleRate * numChannels * 2, true);
+  view.setUint16(32, numChannels * 2, true);
+  view.setUint16(34, 16, true);
+  writeString(view, 36, 'data');
+  view.setUint32(40, samples.length * 2, true);
+
+  floatTo16BitPCM(view, 44, samples);
+
+  return view;
+}
+
+function writeString(view: DataView, offset: number, string: string): void {
+  for (let i = 0; i < string.length; i++) {
+    view.setUint8(offset + i, string.charCodeAt(i));
+  }
+}
+
+function floatTo16BitPCM(output: DataView, offset: number, input: Float32Array): void {
+  for (let i = 0; i < input.length; i++, offset += 2) {
+    const s = Math.max(-1, Math.min(1, input[i]));
+    output.setInt16(offset, s < 0 ? s * 0x8000 : s * 0x7FFF, true);
+  }
+}
